requirements.txt
spotipy==2.24.0
streamlit==1.35.0
pandas==2.2.2
python-dotenv==1.0.1
